+++ examples/Messages.tsx
@@ -18,6 +18,9 @@ const messages = {
         other {#th person}
     }`,
   richtext: 'I have <bold>{num, plural, one {# dog} other {# dogs}}</bold>',
+  richertext:
+    'I have & < &nbsp; <bold>{num, plural, one {# & dog} other {# dogs}}</bold>',
+  unicode: 'Hello\u0020{placeholder}',
 };
 
 const App: React.FC<Props> = () => {
@@ -53,6 +56,25 @@ const App: React.FC<Props> = () => {
         <FormattedMessage
           id="richtext"
           values={{num: 99, bold: (...chunks) => <strong>{chunks}</strong>}}
+        />
+        <br />
+        <FormattedMessage
+          id="richertext"
+          values={{num: 99, bold: (...chunks) => <strong>{chunks}</strong>}}
+        />
+        <br />
+        <FormattedMessage
+          id="random"
+          defaultMessage="I have < &nbsp; <bold>{num, plural, one {# dog} other {# & dogs}}</bold>"
+          values={{num: 99, bold: (...chunks) => <strong>{chunks}</strong>}}
+        />
+        <br />
+        <FormattedMessage id="unicode" values={{placeholder: 'world'}} />
+        <br />
+        <FormattedMessage
+          id="whatever"
+          defaultMessage="Hello\u0020{placeholder}"
+          values={{placeholder: 'world'}}
         />
       </p>
     </IntlProvider>