+++ src/components/injectIntl.tsx
@@ -4,7 +4,8 @@ import * as hoistNonReactStatics_ from 'hoist-non-react-statics';
 // this is to interop with TypeScript since `invariant`
 // does not export a default
 // https://github.com/rollup/rollup/issues/1267
-const hoistNonReactStatics: typeof hoistNonReactStatics_ = require('hoist-non-react-statics');
+const hoistNonReactStatics: typeof hoistNonReactStatics_ =
+  (hoistNonReactStatics_ as any).default || hoistNonReactStatics_;
 import {invariantIntlContext} from '../utils';
 import {IntlShape, Omit} from '../types';
 
@@ -19,9 +20,12 @@ const {Consumer: IntlConsumer, Provider: IntlProvider} = IntlContext;
 export const Provider = IntlProvider;
 export const Context = IntlContext;
 
-export interface Opts<IntlPropName extends string = 'intl'> {
+export interface Opts<
+  IntlPropName extends string = 'intl',
+  ForwardRef extends boolean = false
+> {
   intlPropName?: IntlPropName;
-  forwardRef?: boolean;
+  forwardRef?: ForwardRef;
   enforceContext?: boolean;
 }
 
@@ -34,12 +38,37 @@ export type WithIntlProps<P> = Omit<P, keyof WrappedComponentProps> & {
 };
 
 export default function injectIntl<
-  IntlPropName extends string = 'intl',
+  IntlPropName extends string,
   P extends WrappedComponentProps<IntlPropName> = WrappedComponentProps<any>
 >(
   WrappedComponent: React.ComponentType<P>,
-  options?: Opts<IntlPropName>
-): React.ComponentType<WithIntlProps<P>> & {
+  options?: Opts<IntlPropName, false>
+): React.FC<WithIntlProps<P>> & {
+  WrappedComponent: typeof WrappedComponent;
+};
+export default function injectIntl<
+  IntlPropName extends string = 'intl',
+  P extends WrappedComponentProps<IntlPropName> = WrappedComponentProps<any>,
+  T extends React.ComponentType<P> = any
+>(
+  WrappedComponent: React.ComponentType<P>,
+  options?: Opts<IntlPropName, true>
+): React.ForwardRefExoticComponent<
+  React.PropsWithoutRef<WithIntlProps<P>> & React.RefAttributes<T>
+> & {
+  WrappedComponent: typeof WrappedComponent;
+};
+export default function injectIntl<
+  IntlPropName extends string = 'intl',
+  P extends WrappedComponentProps<IntlPropName> = WrappedComponentProps<any>,
+  ForwardRef extends boolean = false,
+  T extends React.ComponentType<P> = any
+>(
+  WrappedComponent: React.ComponentType<P>,
+  options?: Opts<IntlPropName, ForwardRef>
+): React.ForwardRefExoticComponent<
+  React.PropsWithoutRef<WithIntlProps<P>> & React.RefAttributes<T>
+> & {
   WrappedComponent: typeof WrappedComponent;
 } {
   const {intlPropName = 'intl', forwardRef = false, enforceContext = true} =
@@ -71,7 +100,7 @@ export default function injectIntl<
 
   if (forwardRef) {
     return hoistNonReactStatics(
-      React.forwardRef((props: P, ref) => (
+      React.forwardRef<T, P>((props: P, ref) => (
         <WithIntl {...props} forwardedRef={ref} />
       )),
       WrappedComponent