+++ src/types.ts
@@ -13,6 +13,7 @@ import IntlRelativeTimeFormat, {
   IntlRelativeTimeFormatOptions,
 } from '@formatjs/intl-relativetimeformat';
 import {MessageFormatElement} from 'intl-messageformat-parser';
+import {UnifiedNumberFormatOptions} from '@formatjs/intl-unified-numberformat';
 
 export interface IntlConfig {
   locale: string;
@@ -39,7 +40,7 @@ export type FormatDateOptions = Exclude<
 > &
   CustomFormatConfig;
 export type FormatNumberOptions = Exclude<
-  Intl.NumberFormatOptions,
+  UnifiedNumberFormatOptions,
   'localeMatcher'
 > &
   CustomFormatConfig;
@@ -63,6 +64,14 @@ export interface IntlFormatters {
     value: Parameters<Intl.DateTimeFormat['format']>[0] | string,
     opts?: FormatDateOptions
   ): string;
+  formatDateToParts(
+    value: Parameters<Intl.DateTimeFormat['format']>[0] | string,
+    opts?: FormatDateOptions
+  ): Intl.DateTimeFormatPart[];
+  formatTimeToParts(
+    value: Parameters<Intl.DateTimeFormat['format']>[0] | string,
+    opts?: FormatDateOptions
+  ): Intl.DateTimeFormatPart[];
   formatRelativeTime(
     value: Parameters<IntlRelativeTimeFormat['format']>[0],
     unit?: Parameters<IntlRelativeTimeFormat['format']>[1],
@@ -72,6 +81,10 @@ export interface IntlFormatters {
     value: Parameters<Intl.NumberFormat['format']>[0],
     opts?: FormatNumberOptions
   ): string;
+  formatNumberToParts(
+    value: Parameters<Intl.NumberFormat['format']>[0],
+    opts?: FormatNumberOptions
+  ): Intl.NumberFormatPart[];
   formatPlural(
     value: Parameters<Intl.PluralRules['select']>[0],
     opts?: FormatPluralOptions